packets1: Move `Header` into `header.go`
